+++ packages/gatsby-theme-notes/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.74](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.73...gatsby-theme-notes@1.3.74) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-theme-notes
+
+## [1.3.73](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.72...gatsby-theme-notes@1.3.73) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-theme-notes
+
+## [1.3.72](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.71...gatsby-theme-notes@1.3.72) (2020-06-20)
+
+**Note:** Version bump only for package gatsby-theme-notes
+
+## [1.3.71](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.70...gatsby-theme-notes@1.3.71) (2020-06-19)
+
+**Note:** Version bump only for package gatsby-theme-notes
+
+## [1.3.70](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.69...gatsby-theme-notes@1.3.70) (2020-06-19)
+
+**Note:** Version bump only for package gatsby-theme-notes
+
+## [1.3.69](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.68...gatsby-theme-notes@1.3.69) (2020-06-15)
+
+**Note:** Version bump only for package gatsby-theme-notes
+
 ## [1.3.68](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-notes@1.3.67...gatsby-theme-notes@1.3.68) (2020-06-09)
 
 **Note:** Version bump only for package gatsby-theme-notes